fix tests
- fix docstring + refactor tests
