+++ docs/source/bettertransformer/overview.mdx
@@ -37,6 +37,7 @@ The list of supported model below:
 - [Splinter](https://arxiv.org/abs/2101.00438)
 - [XLMRoberta](https://arxiv.org/abs/1911.02116)
 - [Whisper](https://cdn.openai.com/papers/whisper.pdf)
+- [Wav2Vec2](https://arxiv.org/abs/2006.11477)
 
 Let us know by opening an issue in 🤗 Optimum if you want more models to be supported, or check out the contribution guideline if you want to add it by yourself!
